# Conflicts:
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/activities/charts/WeekStepsChartFragment.java
@@ -105,7 +105,7 @@ public class WeekStepsChartFragment extends AbstractWeekChartFragment {
     @Override
     protected String getBalanceMessage(long balance, int targetValue) {
         if (balance > 0) {
-            final long totalBalance = balance - (targetValue * TOTAL_DAYS);
+            final long totalBalance = balance - ((long)targetValue * TOTAL_DAYS);
             if (totalBalance > 0)
                 return getString(R.string.overstep, Math.abs(totalBalance));
             else
